Remove simple writer module
